requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.4
itsdangerous==2.1.2
python-dotenv==1.0.0
requests==2.31.0
xhtml2pdf==0.2.6   # <-- FIXED
reportlab==3.6.12            # pinned to fix ShowBoundaryValue import
PyMuPDF==1.23.6               # newer, compatible with Python 3.11
google-generativeai==0.5.2

gunicorn==23.0.0
click==8.1.7
MarkupSafe==2.1.5
pyodbc==4.0.35

